// src/components/button/button.enum.ts
export enum EButtonTypes {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  TERTIARY = 'tertiary',
}

export enum EButtonSizes {
  SMALL = 'sm',
  MEDIUM = 'md',
  LARGE = 'lg',
}

export enum EButtonIconPosition {
  LEFT = 'left',
  RIGHT = 'right',
  TOP = 'top',
}

export enum EButtonKinds {
  BUTTON = 'button',
  LINK = 'link',
  BUTTON_AS_LINK = 'buttonAsLink',
  COPY_TO_CLIPBOARD = 'copyToClipboard',
  BUTTON_ICON = 'buttonIcon',
  BUTTON_ICON_MENU = 'buttonIconMenu',
  BUTTON_ICON_MENU_AS_LINK = 'buttonIconMenuAsLink',
}